Move `benchmark_process.rs` -> `process.rs`
